+++ packages/components/src/components/checkbox/checkbox.tsx
@@ -63,7 +63,7 @@ export class Checkbox {
   @Prop() public required?: boolean = false;
 
   /** Marks the checkbox as disabled. */
-  @Prop() public disabled?: boolean = false;
+  @Prop({ mutable: true }) public disabled?: boolean = false;
 
   /** Marks the checkbox as indeterminate. */
   @Prop() public indeterminate?: boolean = false;
@@ -167,6 +167,7 @@ export class Checkbox {
   }
 
   public formDisabledCallback(disabled: boolean): void {
+    // Called when a parent fieldset is disabled or enabled
     this.disabled = disabled;
   }