+++ packages/components/src/components/segmented-control/segmented-control/segmented-control.tsx
@@ -160,6 +160,7 @@ export class SegmentedControl {
   }
 
   public formDisabledCallback(disabled: boolean): void {
+    // Called when a parent fieldset is disabled or enabled
     this.disabled = disabled;
   }