+++ packages/components/src/components/tag-dismissible/tag-dismissible-styles.ts
@@ -1,5 +1,4 @@
 import { borderRadiusSmall, fontSizeTextXSmall, textSmallStyle } from '@porsche-design-system/styles';
-import { spacingStaticXs } from '@porsche-design-system/tokens';
 import {
   addImportantToEachRule,
   colorSchemeStyles,
@@ -10,17 +9,29 @@ import {
   hostHiddenStyles,
   hoverMediaQuery,
   preventFoucOfNestedElementsStyles,
+  SCALING_BASE_VALUE,
 } from '../../styles';
 import { getCss } from '../../utils';
 
 const { primaryColor, frostedColor, contrastHighColor } = colors;
+
+export const cssVarInternalTagDismissibleScaling = '--p-internal-tag-dismissible-scaling';
+export const getScalingVar = (compact: boolean) =>
+  `var(${cssVarInternalTagDismissibleScaling}, ${compact ? 'calc(4 / 13)' : 1})`;
 
 // CSS Variable defined in fontHyphenationStyle
 /**
  * @css-variable {"name": "--p-hyphens", "description": "Sets the CSS `hyphens` property for text elements, controlling whether words can break and hyphenate automatically.", "defaultValue": "auto"}
  */
 
-export const getComponentCss = (hasLabel: boolean): string => {
+export const getComponentCss = (hasLabel: boolean, compact: boolean): string => {
+  const scalingVar = getScalingVar(compact);
+
+  const iconPadding = '4px';
+  const paddingBlock = `calc(${scalingVar} * 0.8125 * ${SCALING_BASE_VALUE} - ${iconPadding}/2)`; // 0.8125 * SCALING_BASE_VALUE corresponds to 13px
+  const paddingInline = `max(calc(${scalingVar} * 0.8125 * ${SCALING_BASE_VALUE} - 1px), 4px)`;
+  const gap = `max(calc(${scalingVar} * 0.75 * ${SCALING_BASE_VALUE}), 2px)`; // 0.5 * SCALING_BASE_VALUE corresponds to 12px
+
   return getCss({
     '@global': {
       ':host': {
@@ -37,9 +48,8 @@ export const getComponentCss = (hasLabel: boolean): string => {
         display: 'flex',
         position: 'relative',
         alignItems: 'center',
-        gap: '12px',
-        minHeight: '54px',
-        padding: '4px 12px',
+        gap,
+        padding: `${hasLabel ? `calc(${paddingBlock} - 6px)` : paddingBlock} ${paddingInline}`,
         borderRadius: borderRadiusSmall,
         cursor: 'pointer',
         background: frostedColor,
@@ -63,7 +73,7 @@ export const getComponentCss = (hasLabel: boolean): string => {
       },
     }),
     icon: {
-      padding: spacingStaticXs,
+      padding: iconPadding,
       marginInlineEnd: '-2px', // compensate white space of svg icon and optimize visual alignment
       transition: getTransition('background-color'),
       borderRadius: borderRadiusSmall,