+++ packages/components/src/styles/select/selected.ts
@@ -1,0 +1,8 @@
+import type { JssStyle } from 'jss';
+
+export const getSelectedSlotJssStyle: JssStyle = {
+  display: 'block',
+  height: '100%',
+  flexGrow: 1,
+  overflow: 'hidden',
+};