+++ packages/shared/scripts/generateCodeExamples.ts
@@ -434,6 +434,15 @@ const codeExamples: CodeSample[] = [
           vue: '/MultiSelectExampleAsyncFilter.vue',
         },
       },
+      {
+        name: 'MultiSelectExampleSelectedSlot',
+        paths: {
+          'vanilla-js': '/multi-select-example-selected-slot.html',
+          angular: '/multi-select-example-selected-slot.component.ts',
+          react: '/MultiSelectExampleSelectedSlot.tsx',
+          vue: '/MultiSelectExampleSelectedSlot.vue',
+        },
+      },
     ],
   },
   {
@@ -529,6 +538,15 @@ const codeExamples: CodeSample[] = [
           angular: '/select-example-async-filter.component.ts',
           react: '/SelectExampleAsyncFilter.tsx',
           vue: '/SelectExampleAsyncFilter.vue',
+        },
+      },
+      {
+        name: 'SelectExampleSelectedSlot',
+        paths: {
+          'vanilla-js': '/select-example-selected-slot.html',
+          angular: '/select-example-selected-slot.component.ts',
+          react: '/SelectExampleSelectedSlot.tsx',
+          vue: '/SelectExampleSelectedSlot.vue',
         },
       },
     ],